# backend/src/Entity/NotificationChannel.php
<?php

declare(strict_types=1);

namespace App\Entity;

use App\Repository\NotificationChannelRepository;
use Doctrine\ORM\Mapping as ORM;
use Symfony\Component\Uid\Uuid;

#[ORM\Entity(repositoryClass: NotificationChannelRepository::class)]
#[ORM\Table(name: 'notification_channels')]
class NotificationChannel
{
    public const TYPE_EMAIL = 'email';
    public const TYPE_TELEGRAM = 'telegram';
    public const TYPE_WEBHOOK = 'webhook';

    #[ORM\Id]
    #[ORM\Column(type: 'uuid')]
    private Uuid $id;

    #[ORM\ManyToOne(targetEntity: Organization::class)]
    #[ORM\JoinColumn(name: 'organization_id', nullable: false, onDelete: 'CASCADE')]
    private Organization $organization;

    #[ORM\Column(length: 30)]
    private string $type;

    #[ORM\Column(length: 255)]
    private string $name;

    /** @var array<string, mixed> */
    #[ORM\Column(name: 'settings_json', type: 'json')]
    private array $settingsJson = [];

    #[ORM\Column]
    private bool $enabled = true;

    #[ORM\Column(name: 'created_at')]
    private \DateTimeImmutable $createdAt;

    #[ORM\Column(name: 'updated_at')]
    private \DateTimeImmutable $updatedAt;

    public function __construct(Organization $organization, string $type, string $name, array $settingsJson = [])
    {
        $this->id = Uuid::v7();
        $this->organization = $organization;
        $this->type = $type;
        $this->name = $name;
        $this->settingsJson = $settingsJson;
        $now = new \DateTimeImmutable();
        $this->createdAt = $now;
        $this->updatedAt = $now;
    }

    public function getId(): Uuid
    {
        return $this->id;
    }

    public function getOrganization(): Organization
    {
        return $this->organization;
    }

    public function getType(): string
    {
        return $this->type;
    }

    public function getName(): string
    {
        return $this->name;
    }

    /** @return array<string, mixed> */
    public function getSettingsJson(): array
    {
        return $this->settingsJson;
    }

    public function isEnabled(): bool
    {
        return $this->enabled;
    }

    public function setEnabled(bool $enabled): self
    {
        $this->enabled = $enabled;
        $this->updatedAt = new \DateTimeImmutable();

        return $this;
    }

    public function getCreatedAt(): \DateTimeImmutable
    {
        return $this->createdAt;
    }
}